+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+-   [Patch] Rewrite `ModalCurtain` component in TypeScript.
+
 ## 12.3.0 - 2020-03-16
 
 ### Added
@@ -11,7 +15,6 @@
 ### Changed
 
 -   [Patch] Rewrite `Modal` component in TypeScript.
--   [Patch] Rewrite `ModalCurtain` component in TypeScript.
 
 ## 12.2.1 - 2020-03-02